started general attrib genes

+++ bugs/genetics/Gene.h
@@ -92,7 +92,7 @@ struct GeneMuscleCommand {
 
 class Gene {
 public:
-	int RID;			// this is unique to this gene, when the gene is mutated, the RID changes
+	unsigned long RID;	// this is unique to this gene, when the gene is mutated, the RID changes
 	gene_type type;		// the type of gene
 	union GeneData {
 		GeneCommand gene_command;
